created datestring generator as a new component for reusing
